+++ instrumentation/opentelemetry-instrumentation-asgi/tests/test_asgi_custom_headers.py
@@ -153,19 +153,19 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
+            "http.request.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.request.header.custom_test_header_2": (
+            "http.request.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
-            "http.request.header.regex_test_header_1": ("Regex Test Value 1",),
-            "http.request.header.regex_test_header_2": (
+            "http.request.header.regex-test-header-1": ("Regex Test Value 1",),
+            "http.request.header.regex-test-header-2": (
                 "RegexTestValue2,RegexTestValue3",
             ),
-            "http.request.header.non_utf8_header": ("Moto Z²",),
-            "http.request.header.moto_z²_non_utf8_header_key": ("Moto Z²",),
-            "http.request.header.my_secret_header": ("[REDACTED]",),
+            "http.request.header.non-utf8-header": ("Moto Z²",),
+            "http.request.header.moto-z²-non-utf8-header-key": ("Moto Z²",),
+            "http.request.header.my-secret-header": ("[REDACTED]",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -183,7 +183,7 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
+            "http.request.header.custom-test-header-1": (
                 "test-header-value-1",
                 "test-header-value-2",
             ),
@@ -202,12 +202,12 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
+            "http.request.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
         }
         not_expected = {
-            "http.request.header.custom_test_header_2": (
+            "http.request.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
         }
@@ -228,21 +228,21 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.response.header.custom_test_header_1": (
+            "http.response.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.response.header.custom_test_header_2": (
+            "http.response.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
-            "http.response.header.my_custom_regex_header_1": (
+            "http.response.header.my-custom-regex-header-1": (
                 "my-custom-regex-value-1,my-custom-regex-value-2",
             ),
-            "http.response.header.my_custom_regex_header_2": (
+            "http.response.header.my-custom-regex-header-2": (
                 "my-custom-regex-value-3,my-custom-regex-value-4",
             ),
-            "http.response.header.my_secret_header": ("[REDACTED]",),
-            "http.response.header.non_utf8_header": ("Moto Z²",),
-            "http.response.header.moto_z²_non_utf8_header_key": ("Moto Z²",),
+            "http.response.header.my-secret-header": ("[REDACTED]",),
+            "http.response.header.non-utf8-header": ("Moto Z²",),
+            "http.response.header.moto-z²-non-utf8-header-key": ("Moto Z²",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -259,7 +259,7 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.response.header.custom_test_header_1": (
+            "http.response.header.custom-test-header-1": (
                 "test-header-value-1",
                 "test-header-value-2",
             ),
@@ -278,7 +278,7 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         not_expected = {
-            "http.response.header.custom_test_header_3": (
+            "http.response.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }
@@ -312,17 +312,17 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.request.header.custom_test_header_1": (
+            "http.request.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.request.header.custom_test_header_2": (
+            "http.request.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
-            "http.request.header.regex_test_header_1": ("Regex Test Value 1",),
-            "http.request.header.regex_test_header_2": (
+            "http.request.header.regex-test-header-1": ("Regex Test Value 1",),
+            "http.request.header.regex-test-header-2": (
                 "RegexTestValue2,RegexTestValue3",
             ),
-            "http.request.header.my_secret_header": ("[REDACTED]",),
+            "http.request.header.my-secret-header": ("[REDACTED]",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -352,7 +352,7 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         not_expected = {
-            "http.request.header.custom_test_header_3": (
+            "http.request.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }
@@ -384,19 +384,19 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         expected = {
-            "http.response.header.custom_test_header_1": (
+            "http.response.header.custom-test-header-1": (
                 "test-header-value-1",
             ),
-            "http.response.header.custom_test_header_2": (
+            "http.response.header.custom-test-header-2": (
                 "test-header-value-2",
             ),
-            "http.response.header.my_custom_regex_header_1": (
+            "http.response.header.my-custom-regex-header-1": (
                 "my-custom-regex-value-1,my-custom-regex-value-2",
             ),
-            "http.response.header.my_custom_regex_header_2": (
+            "http.response.header.my-custom-regex-header-2": (
                 "my-custom-regex-value-3,my-custom-regex-value-4",
             ),
-            "http.response.header.my_secret_header": ("[REDACTED]",),
+            "http.response.header.my-secret-header": ("[REDACTED]",),
         }
         for span in span_list:
             if span.kind == SpanKind.SERVER:
@@ -427,7 +427,7 @@ class TestCustomHeaders(AsyncAsgiTestBase):
         await self.get_all_output()
         span_list = self.exporter.get_finished_spans()
         not_expected = {
-            "http.response.header.custom_test_header_3": (
+            "http.response.header.custom-test-header-3": (
                 "test-header-value-3",
             ),
         }