+++ util/opentelemetry-util-http/src/opentelemetry/util/http/__init__.py
@@ -182,12 +182,12 @@ def remove_url_credentials(url: str) -> str:
 
 
 def normalise_request_header_name(header: str) -> str:
-    key = header.lower().replace("-", "_")
+    key = header.lower()
     return f"http.request.header.{key}"
 
 
 def normalise_response_header_name(header: str) -> str:
-    key = header.lower().replace("-", "_")
+    key = header.lower()
     return f"http.response.header.{key}"